+++ src/sg_component.h
@@ -331,7 +331,8 @@ struct SG_Scene : public SG_Transform {
     X(SG_GEOMETRY_SUZANNE, "SuzanneGeometry")                                          \
     X(SG_GEOMETRY_KNOT, "KnotGeometry")                                                \
     X(SG_GEOMETRY_LINES2D, "LinesGeometry")                                            \
-    X(SG_GEOMETRY_POLYGON, "PolygonGeometry")
+    X(SG_GEOMETRY_POLYGON, "PolygonGeometry")                                          \
+    X(SG_GEOMETRY_POLYHEDRON, "PolyhedronGeometry")
 
 enum SG_GeometryType {
 #define X(name, str) name,
@@ -355,6 +356,7 @@ union SG_GeometryParams {
     CylinderParams cylinder;
     KnotParams knot;
     PolygonParams polygon;
+    PolyhedronType polyhedron;
 };
 
 #define SG_GEOMETRY_MAX_VERTEX_ATTRIBUTES 8
@@ -405,6 +407,7 @@ struct SG_Geometry : SG_Component {
     static void buildCylinder(SG_Geometry* g, CylinderParams* p);
     static void buildKnot(SG_Geometry* g, KnotParams* p);
     static void buildPolygon(SG_Geometry* g, PolygonParams* p);
+    static void buildPolyhedron(SG_Geometry* g, PolyhedronType* p);
 };
 
 // ============================================================================